backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
Django==5.2.6
django-allauth==0.63.3
django-cors-headers==4.3.1
django-filter==24.2
django-oauth2-provider==0.2.6.1
djangorestframework==3.15.1
djangorestframework-oauth==1.1.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2025.10.1
factory-boy==3.3.0
Faker==37.8.0
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
oauthlib==3.3.1
psycopg2-binary==2.9.9
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.0.0
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
requests-oauthlib==1.4.0
rpds-py==0.27.1
shortuuid==1.0.13
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
django-redis==5.4.0
isort==5.13.2
